+++ src/modules/manufacturing/entities/benang-sisa-do.ts
@@ -21,6 +21,10 @@ export class BenangSisaDo extends BaseEntity {
   keterangan?: string;
 
   @Field({ nullable: true })
+  @Column({ name: 'DEPARTMENT' })
+  department?: string;
+
+  @Field({ nullable: true })
   @Column({ name: 'ROW_ID' })
   rowId?: string;
 }